fix(ets): widen parameter `size_bits` to `u16`
`EtsParamDef::size_bits` was a `u8`, which caps a parameter at 255 bits —
31 characters of text. KNX master data ships `String_40Byte`, and vendors
use it: the MDT Push Button Lite gives all eight of its logic description
fields that type. Replicating such a product was therefore impossible,
not merely inconvenient.

The `[u8; N]` path in `get_type_info` already knew about the limit and
rejected anything past it, so a too-wide text parameter failed the build
with a clear message. The macro paths did not. `ets_virtual_params!` and
`__vp_def!` both computed

    size_bits: ($size * 8) as u8,

which wraps silently. The module test device declares

    device_name: String(50) => "Device name" [text_source],

so 50 bytes became 400 bits, and `400 as u8` truncated to 144: the
generated program declared `tTEXT144`, an 18-character field where 50 had
been asked for. Widening the type fixes that as a side effect —
`ModuleApplicationProgram1.mtxml` now emits `tTEXT400`, which is the only
intended output change in this commit.

`TypeInfo::size_bits` and the `ETS_SIZE_BITS` constants generated by
`EtsEnum` and the range-enum macro widen along with it, since they feed
the same field. The `#[ets(bits = N)]` attribute stays a `u8` — it is a
bitfield width within a byte, so widening it would invite nonsense — and
converts at its two use sites.

`TypeNumber::size_in_bit` in the schema stays `u8`: `build_type_def`
already rejects a numeric parameter wider than 63 bits (beyond that
`1i64 << size_bits` overflows), so the narrowing after that guard cannot
truncate. Only text parameters needed the extra range, and `TypeText`
already carried a `u32`.

+++ crates/zweidraehte-device/src/ets.rs
@@ -244,8 +244,14 @@ pub struct EtsParamDef {
     /// part of the struct and use 0.
     pub offset: u16,
 
-    /// Size in bits
-    pub size_bits: u8,
+    /// Size in bits.
+    ///
+    /// Wider than the byte a bit count would suggest because text parameters
+    /// are sized in bits too: KNX master data ships `String_40Byte` (320 bits)
+    /// and vendors use it — the MDT Push Button Lite gives its eight logic
+    /// description fields that type — so a `u8` would cap parameters at 31
+    /// characters and make those products inexpressible.
+    pub size_bits: u16,
 
     /// Bit offset within the byte (0-7)
     pub bit_offset: u8,
@@ -798,7 +804,7 @@ macro_rules! ets_virtual_params {
                         display_name: $display,
                         suffix: None,
                         offset: 0,
-                        size_bits: ($size * 8) as u8,
+                        size_bits: ($size * 8) as u16,
                         bit_offset: 0,
                         param_type: $crate::ets::EtsParamType::String,
                         hidden: false,
